requirements.txt
streamlit
groq
python-dotenv
python-pptx
PyPDF2
python-docx
requests
lxml
